requirements.txt
aiohttp==3.9.3
APScheduler==3.10.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
airbyte-serverless==0.23
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beartype==0.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bokeh==3.3.4
boto3==1.34.64
botocore==1.34.64
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-help-colors==0.9.4
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
effdet==0.4.1
emoji==2.10.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fake-useragent==1.5.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.2.0
feedparser==6.0.2
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-run==0.10.5
google-cloud-secret-manager==2.18.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h3==3.7.7
html2text==2024.2.26
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iopath==0.1.10
ipython==8.22.2
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpath-python==1.0.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_bokeh==4.0.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
layoutparser==0.3.4
linkify-it-py==2.0.3
litellm==1.31.17
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
msal==1.27.0
msg-parser==1.2.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
Office365-REST-Python-Client==2.5.6
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.14.1
opencv-python==4.8.0.76
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
packaging==23.2
pandas==2.2.0
panel==1.3.8
param==2.0.2
parso==0.8.3
pathway==0.8.2
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pexpect==4.9.0
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycurl==7.45.3
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pypandoc==1.12
pyparsing==3.0.9
pypdf==4.0.1
pypdfium2==4.27.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
python-sat==1.8.dev3
pytz==2024.1
pyviz_comms==3.0.1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.3.2
scikit-learn==1.4.1.post1
scipy==1.10.1
selenium==2.53.6
shapely==2.0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlglot==10.6.1
srsly==2.4.8
stack-data==0.6.3
streamlit==1.32.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
timm==0.9.12
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.37.1
triton==2.2.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uc-micro-py==1.0.3
Unidecode==1.3.8
unstructured==0.11.8
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
uritemplate==4.1.1
urllib3==1.26.18
vcrpy==6.0.1
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
widgetsnbextension==4.0.10
wikipedia==1.4.0
Wikipedia-API==0.6.0
wptools==0.4.17
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
